Rename SHA256.hpp/cpp to Crypto.hpp/cpp
